drivers: hwinfo: spi-flash-id: fix hwinfo driver warning
Change hwinfo chip_id from uint32_t to uint8_t* buffer

+++ drivers/flash/soc_flash_w91.c
@@ -301,7 +301,7 @@ static void unpack_flash_w91_get_id(
 	}
 }
 
-int flash_w91_get_id(const struct device *dev, uint32_t *hw_id)
+int flash_w91_get_id(const struct device *dev, uint8_t *hw_id)
 {
 	struct ipc_based_driver *ipc_data = &((struct flash_w91_data *)dev->data)->ipc;
 	uint8_t inst = ((struct flash_w91_config *)dev->config)->instance_id;
